chore: add unit tests and license headers
