+++ plugins/legacy-preset-chart-nvd3/test/utils/tokenize.test.js
@@ -1,0 +1,47 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+import { tokenizeToNumericArray, tokenizeToStringArray } from '../../src/utils/tokenize';
+
+describe('tokenizeToNumericArray', () => {
+  it('evals numeric strings properly', () => {
+    expect(tokenizeToNumericArray('1')).toBe([1]);
+    expect(tokenizeToNumericArray('1,2,3,4')).toBe([1, 2, 3, 4]);
+    expect(tokenizeToNumericArray('   1, 2,   3,    4 ')).toBe([1, 2, 3, 4]);
+  });
+
+  it('evals empty string to null', () => {
+    expect(tokenizeToNumericArray('')).toBe(null);
+  });
+
+  it('throws error on incorrect string', () => {
+    expect(tokenizeToNumericArray('qwerty,1,2,3')).toBe(new Error());
+  });
+});
+
+describe('tokenizeToStringArray', () => {
+  it('evals numeric strings properly', () => {
+    expect(tokenizeToNumericArray('a')).toBe(['a']);
+    expect(tokenizeToNumericArray('1,2,3,4')).toBe(['1', '2', '3', '4']);
+    expect(tokenizeToNumericArray('1,a,3, bc ,d')).toBe(['1', 'a', '3', 'bc', 'd']);
+  });
+
+  it('evals empty string to null', () => {
+    expect(tokenizeToNumericArray('')).toBe(null);
+  });
+});
